Add AUTHORS file
The major purpose of this file is to clarify who the current
single copyright holder is (Mesosphere, Inc) and to provide
a place for calling out future additional copyright holders.

Another purpose of this file is to provide a place for giving
credit to a number of people who have contributed to this
software project given that the public commit log of the
open-source release is truncated. The (ir)relevance of this
is concisely described by [1], quote:

"""
At some point in the life of an open source project, a
contributor may ask to be included in the list of copyright
holders. By default, contributors (or their employers) hold
copyright on their contributions, whether or not we note
that publicly, and the commit log is the most accurate
record of their work. But it’s reasonable to want special
public credit for your work.
"""

[1] https://opensource.google.com/docs/releasing/authors